Rename nnlearner to neuralnet
NeuralNet really is just supposed to be a neural network / function
approximator.
